Stop auto-exporting the livegrep results and reuse buffers
Unload scratch buffers instead of wiping them allows reusing the same
buffer for subsequent tasks.

Populating the quickfix list causes buffers to be created for the
entries. This can clutter the bufferlist (though they are unlisted)
given the potentially very large number of matches.

The X normal mode mapping was added for manually exporting the results.