Filter quickfix lists
Mark an entry in the quickfix list with <Tab> and press | or !| to create a
new quickfix list from the selected or unselected items, respectively.
<S-Tab> clears the selection.

By default, the filtered list is placed after the current list, and all
following lists are cleared. To place it at the end, add a positive
count to the filter mapping, i.e. 1| or 1!|.

Other mappings:
!! ->  clear all but the current list
!$ ->  add a duplicate of the current list to the end of the stack.