Add Tmux session information in tabline
Display basic session information in the Vim tabline when Tmux is not
showing its own status bar. Click it to choose from a tree.

This commit also simplifies the Tmux prompt section so that the Vim
tabline and the console prompt look the same. All the special-casing was
more confusing than helpful.